pyproject.toml
[project]
name = "asta-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=0.3.16",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.5",
    "pytest>=9.0.2",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.2.1",
    "rich>=13.0.0",
    "sqlmodel>=0.0.24",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
]
